backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pydicom==2.4.4
pillow==10.2.0
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.27.0
openai==1.14.0
python-jose==3.3.0
numpy==1.26.3
requests==2.31.0
inference-sdk==0.1.4
# DICOM processing dependencies
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
gdcm==3.0.20
